feat: Add color and formatting to output
This commit makes three changes to the output styling.

A color with a random hue is selected for the bar.

The tooltip reports the name of the hovered item and formats the time
with only two decimal places.

The more deeply nested elements have names that include their parent
elements. E.g. "chore.clean.dishes" instead of "dishes".