feat: Update parsing algorithm
The previous algorithm didn't really take the time window into account.
For instance, say the time window is 60 minutes. If you had an event
that started 70 minutes ago and lasted 20 minutes, then the total amount
of time that is registered will be 70 minutes because the algorithm
didn't cut that 10 minutes off.

This algorithm respects the window.