fix: Raise all text nodes to top
This fixes an issue where component bodies would cover the text of
labels of other components due to the order the SVG is drawn in.

Note that this adds a D3 dependency. The only reason I thought this was
acceptable is that I would like to use D3 for other things in the
future.