feat: The timeline uses the latest ideal map
I went back and forth on how to handle the timeline. Some background:

The points values in the timeline are calculated using the ideal values,
so we need to know which ideal map to use: the ideal map at the time the
band in the timeline was created, or the current ideal map.

This will be controversial, but I selected the current ideal map. I
desire a consistent measurement, and even though it seems a bit unfair
to grade older bands based on newer criteria, I think it is valuable to
have a single criteria to judge all bands.

Or something.