feat: Always show all ideal categories
Previously, I was getting the keys from the tree that stored that day's
activities. This meant that the major categories wouldn't show up in the
statistics table until it was initiated.

I wanted to clean up this function anyway, so I just rewrote it using
the keys from the ideal hashmap instead.