feat: Change styling for tooltip when close to border
Previously, the tooltip couldn't really be viewed when the mouse was
close to the edge of the screen because the content was too wide to be
shown to the right of the cursor.

This change shifts the position of the tooltip to the left if it is
within 100 pixels of the right edge of the screen, fixing this issue.